Merged revisions 137780 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r137780 | seanbright | 2008-08-14 10:03:03 -0500 (Thu, 14 Aug 2008) | 8 lines

If we detect that we are no longer connected, try to reconnect a few times
before giving up.  This relies on the timeout settings in the freetds.conf
file and, unfortunately, on a recent version of FreeTDS (0.82 or newer).

I either need to change the current execs to be non-blocking (which I do
not want to do) or we have to force people to run with the latest and
greatest of FreeTDS.  I'm on the fence...

........