Merged revisions 140559,140563,140566 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r140559 | seanbright | 2008-09-02 09:41:41 -0500 (Tue, 02 Sep 2008) | 4 lines

When a call is rejected because of call-limit, the channel driver is behaving
as expected, so we shouldn't report it as an error.  Change to LOG_NOTICE
instead.

........
r140563 | russell | 2008-09-02 10:09:20 -0500 (Tue, 02 Sep 2008) | 3 lines

I'm not sure how these files got to trunk (probably my fault), but they
should not be here

........
r140566 | russell | 2008-09-02 10:11:53 -0500 (Tue, 02 Sep 2008) | 2 lines

Update instructions for getting libresample

........