Merged revisions 141217,141267 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r141217 | mmichelson | 2008-09-05 11:00:24 -0500 (Fri, 05 Sep 2008) | 14 lines

Commit 140417 had a logic flaw in it which
caused port 5060 to always be used when dialing
a peer if no explicit port was specified. This
broke the behavior of implicitly using the port
from which the peer registered if no port is
specified. This commit fixes the logic flaw.

(closes issue #13424)
Reported by: mdu113
Patches:
      13424.patch uploaded by putnopvut (license 60)
Tested by: mdu113


........
r141267 | mmichelson | 2008-09-05 11:35:54 -0500 (Fri, 05 Sep 2008) | 11 lines

Since greetings are not stored in IMAP, we should
not be DISPOSE'ing of them the same way we do with
other messages.

(closes issue #13414)
Reported by: mthomasslo
Patches:
      13414v2.patch uploaded by putnopvut (license 60)
Tested by: mthomasslo


........