Merged revisions 143226 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
................
r143226 | jpeeler | 2008-09-16 11:35:32 -0500 (Tue, 16 Sep 2008) | 17 lines

Blocked revisions 143204 via svnmerge

........
r143204 | jpeeler | 2008-09-16 11:14:25 -0500 (Tue, 16 Sep 2008) | 10 lines

(closes issue #13465)
(closes issue #13425)
Reported by: mdu113
Patches:
      bug13425.diff uploaded by jpeeler (license 325)
Tested by: mdu113

The main change here was to masquerade the channel if the channel that was to be parked was running a PBX on it. The PBX thread can then maintain full control of the channel (the zombie) as it expects to while allowing the parking thread full control of the real (parked) channel.


........

................