First small batch of changes in reference to issue 5014.  We generate a magic call-id to send in our dialog-info messages, which the SNOM will turn around and send back as the Replaces header of a fresh INVITE.  That part works, now we just need to actually pick up the extension.
As discussed with russellb and kpfleming on #asterisk-dev, some of the changes
made in pkempgen-pickup-60212.patch in reference to cid name/num I've left out,
since an extension could be in a ringing state for any number of reasons, and
we can't assume that information will even be useful to the user.

Thanks to russellb and kpfleming for talking me through this stuff.