Add better and more explicit privacy support for the P-Asserted-Identity header.
Parsing was way off since I misread the RFC and didn't
realize that Privacy was a separate header from
P-Asserted-Identity. The privacy should be honored
better, however there is still no trust mechanic 
built into chan_sip (or the rest of Asterisk) for
that matter.

Also, on the advice of an implementer, instead of
transmitting no P-Asserted-Identity when the information
should be private, use the URI "anonymous@anonymous.invalid"