Merged revisions 146883,146920,146923,146925,146928 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r146883 | jpeeler | 2008-10-06 17:26:25 -0500 (Mon, 06 Oct 2008) | 12 lines

This commit squashes together three commits because the wrong approach was originally used. (One of the commits was only one line.)

1) r143204:
The main change here was to masquerade the channel if the channel that was to be parked was running a PBX on it. The PBX thread can then maintain full control of the channel (the zombie) as it expects to while allowing the parking thread full control of the real (parked) channel.

2) r143270:
Changed park_call_full to hold the parkinglot lock a little longer, which protects the parkeduser struct from being freed out from underneath. Made sure that the parking extension is added to the parking context while holding the lock thereby ensuring that there are no spurious warnings from removal attempts when a hangup occurs while the parking lot is being announced.

3) r143475: (the one liner)
compare peer and chan instead of looking at the parked user (pu), which could have possibly already have been freed by the parking thread


........
r146920 | jpeeler | 2008-10-06 17:59:58 -0500 (Mon, 06 Oct 2008) | 2 lines

Mvanbaak said this was needed to compile on OpenBSD, so put it in the OpenBSD section.

........
r146923 | jpeeler | 2008-10-06 18:08:21 -0500 (Mon, 06 Oct 2008) | 3 lines

Similar to r143204, masquerade the channel in the case of Park being called from AGI.


........
r146925 | mvanbaak | 2008-10-06 18:14:33 -0500 (Mon, 06 Oct 2008) | 8 lines

All ODBC parts can now use either unixodbc or iodbc.
This allows for the ODBC parts to work on OpenBSD as well.

99.99% of the work is done by seanbright (bow, bow) and I actually
did nothing but test and yell at him that it still didn't work :)

Thanks for helping out !

........
r146928 | tilghman | 2008-10-06 18:21:02 -0500 (Mon, 06 Oct 2008) | 3 lines

Update documentation; AST_THREADSTORAGE() in trunk only takes a single
argument.

........