Merged revisions 148120,148144,148160 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r148120 | tilghman | 2008-10-09 18:25:53 -0500 (Thu, 09 Oct 2008) | 6 lines

Fix example schema
(closes issue #12860)
 Reported by: flyn
 Patches: 
       res_ldap.conf.patch uploaded by flyn (license 503)

........
r148144 | mmichelson | 2008-10-09 18:30:47 -0500 (Thu, 09 Oct 2008) | 10 lines

Read the callerid in the correct order and make sure to
read the Urgent flag value from the IMAP headers.

(closes issue #13652)
Reported by: jaroth
Patches:
      imapheaders.patch uploaded by jaroth (license 50)



........
r148160 | mmichelson | 2008-10-09 18:54:59 -0500 (Thu, 09 Oct 2008) | 14 lines

The priority was unnecessary for the manager atxfer, so it has
been removed. Furthermore, now we actually use the Context argument
passed to set the transfer context and don't error out if no
context is specified.

This addresses the actual problems outlined in issue 12158. Regarding
the other points brought up, regarding the inability to not transfer
to extensions which cannot be represented by DTMF, it is not enough of
a constraint that it is worth attempting to rework the feature.

(closes issue #12158)
Reported by: davidw


........