Merged revisions 148373,148376 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r148373 | mmichelson | 2008-10-10 16:18:10 -0500 (Fri, 10 Oct 2008) | 8 lines

Make sure that the inUse and inRinging fields for 
a sip peer cannot go below zero. This is a regression
from 1.4 and so it will be applied to 1.6.0 as well.

(closes issue #13668)
Reported by: mjc


........
r148376 | mmichelson | 2008-10-10 16:21:45 -0500 (Fri, 10 Oct 2008) | 13 lines

The logic used when checking a peer got changed subtly
in the "kill the user" commit and caused calls relying
on the insecure setting to not work properly. I changed
for finding a peer back to how it was prior to that 
commit.

(closes issue #13644)
Reported by: pj
Patches:
      13644_trunkv2.patch uploaded by putnopvut (license 60)
Tested by: pj


........