Merged revisions 148570 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r148570 | tilghman | 2008-10-13 19:08:52 -0500 (Mon, 13 Oct 2008) | 10 lines

Merge realtime_update2 branch, which adds a new realtime API call named
'update2', which permits updates which match across multiple columns, instead
of requiring all tables to have a single unique identifier.  All of the other
API calls with the exception of 'update' already had the ability to match on
multiple fields, so it was a missing and very desireable feature that an API
call implementing an update should have this, too.

This does not change any outward performance of Asterisk, but it should make
life easier for application developers who use the RealTime framework.

........