First pass at making sure that connected line logic is working properly for parallel dialing and ringall queues, as well as fixing logic used with local channels.
For dialing multiple destinations, we handle connected line
updates by locally updating the outgoing structure in app_dial
or app_queue and then applying this change once someone actually
answers. There are some XXX comments indicating that I need to
fix the answer logic up somewhat considering that there is 
redundancy in there.

Local channels are another beast. Since a connected line control
frame is only meant to *update* the channel receiving the indication,
it does not contain all pieces of information necessary. Therefore,
when a local channel attempts to use its default action of queuing
the data to the other side of the local channel, it means incomplete
information is reaching the original outgoing channel. This has been
fixed by using the cumulatively collected connected line information
present in the local channel when queueing the update instead of
just what was in the locally indicated frame. Note that there is a
XXX comment here since I'm not sure if there may actually be potential
for deadlock here. I will need to investigate further.

As said before, this is a first pass, more tweaks will happen...tomorrow!