Work out the logic behind connected line information when multiple destinations are dialed.
So, here's how it works. In app_dial and app_queue, there
is a structure for representing an outgoing call (app_queue
calls this structure a callattempt, for instance). When we
allocate our call attempts (before ever requesting channels
or calling), we first check to see if the dialing channel
has had a connected line number set on it. If it is, we take
this as an indication that some information has been set before
the dial has taken place and that we'd like to use this if
no updating information comes in. Cool.

Later, when we request the outgoing channel, we check to see
if there is a callerid number set. If there is not, then we set
some sort of flag to indicate that the callerid from the far
end sucks.

When new connected line information comes in, we update the
callattempt's connected line structure with the new info we
just got.

When someone finally answers the phone, our first preference is
to use whatever connected line information we have saved in the
callattempt's connected line element. If the number portion is
NULL, then we take this to mean that we never set any connected
line information there, so we fall back to the callerid for the
channel. If that sucks, then we send no connected line update on
the channel.

The most questionable thing that's happening right now is that
we set outgoing connected line based on the fact that the number
portion is non-NULL, which may be fallacious. More tests will help
to confirm all of this.