A set of changes backported from the issue8824 branch, specifically dealing with connected line logic when multiple parties are dialed and when local channels are dialed.
To be more specific, this deals with the fact that when
dialing multiple parties, connected line updates will
be flying in from all directions. So now, we cache these
updates and associate them with a destination dialed.
When someone answers, we then send the appropriate
connected line update.

As far as local channels go, the problem is that if
someone dialed a local channel, then the outgoing
leg of the local channel would have connected line
updates happen, but would not have complete updates
forwarded to the original calling channel. I modified
local_indicate to queue the connected line information
from the outgoing local channel instead of just the
data in the control frame.