Merged revisions 153365 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r153365 | russell | 2008-11-01 16:10:07 -0500 (Sat, 01 Nov 2008) | 11 lines

Merge changes from team/group/appdocsxml

This commit introduces the first phase of an effort to manage documentation of the
interfaces in Asterisk in an XML format.  Currently, a new format is available for
applications and dialplan functions.  A good number of conversions to the new format
are also included.

For more information, see the following message to asterisk-dev:

http://lists.digium.com/pipermail/asterisk-dev/2008-October/034968.html

........