This introduces a new option for sip.conf, rpid_immediate.
By setting this option to "yes," any connected line updates
which occur during call setup will be immediately sent as
provisional messages to the caller's channel. By setting this
to "no" the information will instead be collected and sent out
in the next provisional response that is sent.

The option defaults to being set "off"

Closes AST-135