Merged revisions 156817 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
................
r156817 | mmichelson | 2008-11-14 09:20:03 -0600 (Fri, 14 Nov 2008) | 18 lines

Merged revisions 156816 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r156816 | mmichelson | 2008-11-14 09:18:59 -0600 (Fri, 14 Nov 2008) | 10 lines

If the prompt to reenter a voicemail password timed out, it
resulted in the password not being saved, even if the input matched
what you gave when first prompted to enter a new password. This is
because the return value of ast_readstring was checked, but not checked
properly.

This bug was discovered by Jared Smith during an Asterisk training course.
Thanks for reporting it!


........

................