Merged revisions 158754,158756 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r158754 | seanbright | 2008-11-22 21:30:46 -0600 (Sat, 22 Nov 2008) | 3 lines

No need to use a separate structure for this since we can just pass
our sip_pvt pointer in directly.

........
r158756 | seanbright | 2008-11-22 21:36:52 -0600 (Sat, 22 Nov 2008) | 6 lines

If you enabled 'notifycid' one of the limitations is that the calling channel
is only found if it dialed the extension that was subscribed to.  You can now
specify 'ignore-context' for the 'notifycid' option in sip.conf which will, as
it's value implies, ignore the current context of the caller when doing the
lookup.

........