Merged revisions 158959 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r158959 | seanbright | 2008-11-24 19:01:49 -0600 (Mon, 24 Nov 2008) | 8 lines

This is basically a complete rollback of r155401, as it was determined that
it would be best to maintain API compatibility.  Instead, this commit introduces
ao2_callback_data() which is functionally identical to ao2_callback() except
that it allows you to pass arbitrary data to the callback.

Reviewed by Mark Michelson via ReviewBoard:
	http://reviewboard.digium.com/r/64

........