Merged revisions 159734 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r159734 | mvanbaak | 2008-11-28 08:20:11 -0600 (Fri, 28 Nov 2008) | 8 lines

Make res_config_ldap compile with the official OpenLDAP 2.3.X versions.
They removed the LDAP_DEPRECATED define from their source and since we are using a couple
of deprecated function calls we should define it with a CFLAG.

Tested by me on OpenBSD 4.4 and snuff-home on Linux to make sure everything keeps compiling.
It shouldn't break, we only define the LDAP_DEPRECATED with this which is what 
all 2.2.X and older versions of OpenLDAP did in their own tree.

........